refactor: move helper method to utils
