+++ packages/auth0-acul-js/src/screens/reset-password/index.ts
@@ -1,7 +1,7 @@
 import { ScreenIds } from '../../constants';
-import coreValidatePassword from '../../helpers/validatePassword';
 import { BaseContext } from '../../models/base-context';
 import { FormHandler } from '../../utils/form-handler';
+import coreValidatePassword from '../../utils/validatePassword';
 
 import { ScreenOverride } from './screen-override';
 import { TransactionOverride } from './transaction-override';
@@ -59,32 +59,32 @@ export default class ResetPassword extends BaseContext implements ResetPasswordM
    * // result => { valid: true, errors: [] }
    */
 
-   /**
-  * Validates a password string against the current transaction's password policy.
-  *
-  * This method retrieves the password policy from the current transaction context
-  * and delegates the actual validation to `coreValidatePassword`.
-  *
-  * It returns an array of validation results, each containing:
-  * - `code`: the identifier of the password rule,
-  * - `policy`: a user-friendly description of the rule,
-  * - `isValid`: boolean indicating if the password passed that rule.
-  *
-  * @param {string} password - The password string to validate.
-  * @returns {PasswordRuleValidation[]} An array of rule validation results.
-  *
-  * @example
-  * ```ts
-  * const resetPassword = new ResetPassword();
-  * const validationResults = resetPassword.validatePassword('MyP@ssw0rd!');
-  * console.log(validationResults);
-  * // [
-  * //   { code: 'password-policy-length-at-least', policy: 'At least 12 characters', isValid: false },
-  * //   { code: 'password-policy-lower-case', policy: 'Lowercase letters (a-z)', isValid: true },
-  * //   ...
-  * // ]
-  * ```
-  */
+  /**
+ * Validates a password string against the current transaction's password policy.
+ *
+ * This method retrieves the password policy from the current transaction context
+ * and delegates the actual validation to `coreValidatePassword`.
+ *
+ * It returns an array of validation results, each containing:
+ * - `code`: the identifier of the password rule,
+ * - `policy`: a user-friendly description of the rule,
+ * - `isValid`: boolean indicating if the password passed that rule.
+ *
+ * @param {string} password - The password string to validate.
+ * @returns {PasswordRuleValidation[]} An array of rule validation results.
+ *
+ * @example
+ * ```ts
+ * const resetPassword = new ResetPassword();
+ * const validationResults = resetPassword.validatePassword('MyP@ssw0rd!');
+ * console.log(validationResults);
+ * // [
+ * //   { code: 'password-policy-length-at-least', policy: 'At least 12 characters', isValid: false },
+ * //   { code: 'password-policy-lower-case', policy: 'Lowercase letters (a-z)', isValid: true },
+ * //   ...
+ * // ]
+ * ```
+ */
   validatePassword(password: string): PasswordRuleValidation[] {
     const passwordPolicy = this.transaction?.passwordPolicy;
     return coreValidatePassword(password, passwordPolicy);
