+++ packages/auth0-acul-js/src/screens/signup-password/index.ts
@@ -1,7 +1,7 @@
 import { ScreenIds } from '../../constants';
-import coreValidatePassword from '../../helpers/validatePassword';
 import { BaseContext } from '../../models/base-context';
 import { FormHandler } from '../../utils/form-handler';
+import coreValidatePassword from '../../utils/validatePassword';
 
 import { ScreenOverride } from './screen-override';
 import { TransactionOverride } from './transaction-override';
