+++ packages/auth0-acul-js/src/utils/validatePassword.ts
