+++ packages/auth0-acul-js/tests/unit/utils/validateUsername.test.ts
@@ -1,4 +1,4 @@
-import validateUsername from '../../../src/helpers/validateUsername';
+import validateUsername from '../../../src/utils/validateUsername';
 import type { UsernamePolicy } from '../../../interfaces/models/transaction';
 
 // Helper to extract failed error codes
